Product/server/requirements.txt
alabaster==0.7.12
apipkg==1.4
astroid==2.1.0
atomicwrites==1.3.0
attrs==19.3.0
Babel==2.8.0
beautifulsoup4==4.8.2
bleach==3.1.0
blis==0.2.4
boto==2.49.0
boto3==1.11.9
botocore==1.14.9
bs4==0.0.1
catalogue==1.0.0
certifi==2017.7.27.1
chardet==3.0.4
Click==7.0
coverage==4.0.3
cymem==2.0.3
Cython==0.29.12
datefinder-lexpredict==0.6.2
dateparser==0.7.0
docutils==0.15.2
elasticsearch==5.4.0
en-core-web-sm==2.1.0
execnet==1.5.0
Flask==1.1.1
gensim==3.4.0
idna==2.6
imagesize==1.2.0
importlib-metadata==1.5.0
isort==4.2.15
itsdangerous==1.1.0
jellyfish==0.5.6
Jinja2==2.11.1
jmespath==0.9.4
joblib==0.14.1
jsonschema==2.6.0
lazy-object-proxy==1.3.1
lxml==4.2.0
MarkupSafe==1.1.1
mccabe==0.6.1
memory-profiler==0.47
more-itertools==8.2.0
murmurhash==1.0.2
neuralcoref==4.0
nltk==3.4.5
nose==1.3.7
num2words==0.5.7
numpy==1.15.1
packaging==20.1
pandas==0.23.4
pep8==1.7.0
pkginfo==1.5.0.1
plac==0.9.6
pluggy==0.12.0
preshed==2.0.1
psutil==5.4.0
py==1.7.0
pycountry==18.5.26
Pygments==2.5.2
pylint==2.2.2
pyparsing==2.4.6
pytest==5.2.1
pytest-cache==1.0
pytest-cov==2.5.1
pytest-forked==1.1.3
pytest-pep8==1.0.6
pytest-pylint==0.14.1
pytest-xdist==1.30.0
python-coveralls==2.9.1
python-dateutil==2.6.1
python-docx==0.8.10
pytz==2017.2
PyYAML==5.1.2
readme-renderer==24.0
regex==2017.9.23
reporters-db==1.0.12.1
requests==2.22.0
requests-toolbelt==0.9.1
s3transfer==0.3.2
scikit-learn==0.21.3
scipy==1.0.0
six==1.11.0
sklearn==0.0
smart-open==1.9.0
snowballstemmer==2.0.0
soupsieve==1.9.5
spacy==2.1.0
Sphinx==2.1.2
sphinxcontrib-applehelp==1.0.1
sphinxcontrib-devhelp==1.0.1
sphinxcontrib-htmlhelp==1.0.2
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.2
sphinxcontrib-serializinghtml==1.1.3
srsly==1.0.1
thinc==7.0.8
tika==1.15
tqdm==4.42.0
twine==1.13.0
tzlocal==2.0.0
Unidecode==0.4.21
urllib3==1.25.3
us==1.0.0
wasabi==0.6.0
wcwidth==0.1.8
webencodings==0.5.1
Werkzeug==1.0.0
wrapt==1.10.11
zipp==2.1.0
